Remove generation of workflow.org
NOTE: Used to call this method in main and it would create a nice workflow.org
in the assets/org-include directory but the format was very cumbersome as I
could not see what I needed to see and because half the data came from
airtable it became hard to annotate it my style. So basically deprecating this
but still keeping the function for a while because the utility itself is
useful and might need somewhere else

NOTE: When removing this method, might as well extract it out into something
solid rather than completely getting rid of it

NOTE: This also removes the following from the makefile. The airtable table
will stay for a while.
airtable-export ${TEMP_DIR} appHCIWcQJnvFYG2n workflows --json